Fix ListManager external window design
